Use existing class decl summary comment if one exists.
Added test to keep existing summary documentation comment if one exists for a class declaration. Added code to retrieve the summary comment and add it back into the source code.

We don't want to overwrite existing comment with the default. That would just undo work spent on documenting the class declaration.