Add full name of a base class or interface in seealso comment.
For a class declaration with a generic base class or generic interface, provide the full name of the base class or interface in the seealso documentation comment. For example, for an interface of IEnumerable<int> where IEnumerable is really System.Collections.Generic.IEnumerable<out T>, add <seealso cref="System.Collections.Generic.IEnumerable`1"/>

It was also necessary to ensure that any empty lines preceeding the class declaration are preserved.