Use semantic model to get base class information.
To gather information about base class and interfaces, it was necessary to use the semantic model. Changes were also made to better support DoxyPress output. This required changing the seealso output from IEnumerable`1 to IEnumerable.

Added test for class with interface defined in the code as opposed to defined in one of the using directives. Modified a number of tests to satisfy the new seealso output.